WIP: FIX examples
