+++ gmx2qmmm/operations/nma_3N_6dof.py
